추가 리뷰사항 반영

+++ infra/social/src/main/java/com/dragons/client/GoogleTokenResponse.java
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 public record GoogleTokenResponse(
     @JsonProperty("access_token") String accessToken,
-    @JsonProperty("expires_in") Integer expiresIn,
+    @JsonProperty("expires_in") Long expiresIn,
     @JsonProperty("token_type") String tokenType,
     @JsonProperty("scope") String scope,
     @JsonProperty("refresh_token") String refreshToken
